Deploy 394

+++ Backend_UMR_Work_Program/DataModels/WP_HSE_FATALITIES_accident_statistic_table.cs
@@ -8,8 +8,8 @@ namespace Backend_UMR_Work_Program.Models
         public string? CompanyName { get; set; }
         public string? Year_of_WP { get; set; }
         public string? Fatalities_Type { get; set; }
-        public int? Fatalities_Proposed_year { get; set; }
-        public int? Fatalities_Current_year { get; set; }
+        public decimal? Fatalities_Proposed_year { get; set; }
+        public decimal? Fatalities_Current_year { get; set; }
         public string? type_of_incidence { get; set; }
     }
 }
